refactor(auth): process -> handler에 역할배정
